Move dithering to a new file. Add Sierra dithering algoritms
